chore: Move `BottomActions` to `components` folder
